+++ db/pvault.sql
@@ -299,6 +299,7 @@ CREATE TABLE `settings` (
   `pv_maker` int(11) DEFAULT 0,
   `pv_maker_host` varchar(255) COLLATE utf8_bin DEFAULT NULL,
   `qStep` INT(5) NOT NULL DEFAULT '2',
+  `pubchem_view` VARCHAR(4) NOT NULL DEFAULT '2d', 
   `defCatClass` VARCHAR(255) NOT NULL DEFAULT 'cat4' 
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;