+++ db/updates/update_2.5-2.6.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `settings` ADD `pubchem_view` VARCHAR(4) NOT NULL DEFAULT '2d' AFTER `chkVersion`;