+++ db/updates/update_2.5-2.6.sql
@@ -1,2 +1,2 @@
 ALTER TABLE `settings` ADD `pubchem_view` VARCHAR(4) NOT NULL DEFAULT '2d' AFTER `chkVersion`;
-ALTER TABLE `ingredients` ADD `tradingName` VARCHAR(255) NULL AFTER `name`;
+ALTER TABLE `ingredients` ADD `INCI` VARCHAR(255) NULL AFTER `name`;