+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ### Version 2.6
 - ADD: Option to select between 2d and 3d view in PubChem intergration
+- FIX: Issue preventing printing back box label
 
 ### Version 2.5
 - FIX: Docker helper tag (set to :latest)