+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 ### Version 2.6
+- ADD: Trading name in ingredients
 - ADD: Option to select between 2d and 3d view in PubChem intergration
 - FIX: Issue preventing printing back box label
 - FIX: Docker helper minor update