+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Version 2.6
 - ADD: Trading name in ingredients
 - ADD: Option to select between 2d and 3d view in PubChem intergration
+- ADD: SG help reference
 - FIX: Issue preventing printing back box label
 - FIX: Docker helper minor update
 - FIX: A bug preventing showing fromula image in todo drop-down menu