+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - FIX: Issue preventing printing back box label
 - FIX: Docker helper minor update
 - FIX: A bug preventing showing fromula image in todo drop-down menu
+- FIX: INCI name order
 - UPD: Rename insights page to statistics
 
 ### Version 2.5